updating Linux build for new volume API
git-svn-id: http://audiere.svn.sourceforge.net/svnroot/audiere/trunk@235 cb5fa4c3-b70d-0410-b4f6-90f5898e5689

+++ src/device_al.h
@@ -42,8 +42,8 @@ namespace audiere {
     void stop();
     void reset();
     bool isPlaying();
-    void setVolume(int volume);
-    int  getVolume();
+    void  setVolume(float volume);
+    float getVolume();
 
   private:
     ALOutputStream(
@@ -79,7 +79,7 @@ namespace audiere {
 
     int m_buffer_length;  // in samples
     bool m_is_playing;
-    int m_volume;
+    float m_volume;
 
     friend class ALAudioDevice;
   };
